% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session2.script.R, R/classic.R
\name{input_classic}
\alias{input_classic}
\title{Classical R appraoches to data analysis}
\usage{
input_classic(pdata_file, exprs_file)

input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the pheno data file.}

\item{exprs_file}{character(1) The path to the expression file.}

\item{pdata_file}{character(1) The path to the pheno data file.}

\item{exprs_file}{character(1) The path to the expression file.}
}
\value{
A data.frame() containing samples as rows and phenotype data and
expression data as columns.

A data.frame() containing samples as rows and phenotype data and
expression data as columns.
}
\description{
The functions here illustrate how an "old school" R user might approach data
management and analysis

The functions here illustrate how an "old school" R user might approach data
management and analysis
}
\details{
There are two files.  They are from a microarray experiment. The
first file ALL-phenoData.csv describes the samples.  The second file,
ALL-expression.csv are the normalized expression values for 12k probesets
across 128 samples. The files originally came from sheets in excel, and
were exported from Excel using export-to-csv.

There are two files.  They are from a microarray experiment. The
first file ALL-phenoData.csv describes the samples.  The second file,
ALL-expression.csv are the normalized expression values for 12k probesets
across 128 samples. The files originally came from sheets in excel, and
were exported from Excel using export-to-csv.
}

